% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sm_optim_rmse}
\alias{sm_optim_rmse}
\alias{sm_optim_ssq}
\title{Objective Functions}
\usage{
sm_optim_rmse(x, y, ...)

sm_optim_ssq(x, y, ...)
}
\arguments{
\item{x}{numeric. Vector of "observed" values.}

\item{y}{numeric. Vector of "predicted" values.}

\item{...}{Additional arguments to statistics functions \code{mean()}, \code{sum()} etc. such as \code{na.rm}.}
}
\value{
numeric
}
\description{
Objective Functions
}
\examples{
sm_optim_rmse(1:10, runif(10))
sm_optim_ssq(1:10, runif(10))
}
