% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_gradational}
\alias{sm_shape_gradational}
\title{Gradational Shape Function}
\usage{
sm_shape_gradational(x, xlim, ascending = TRUE, type = "linear")
}
\arguments{
\item{x}{numeric. Depth vector.}

\item{xlim}{numeric. Parameters: c(start_value, end_value). Values in the range 0 to 1.}

\item{ascending}{logical. If TRUE, increases from start to end. If FALSE, decreases.}

\item{type}{character. "linear" (default) or "smooth" (polynomial).}
}
\value{
numeric vector in the range 0 to 1 representing the shape.
}
\description{
Models a linear or smooth change with depth. Useful for steady accumulation processes.
}
\examples{
plot(sm_shape_gradational(0:100, c(0.2, 0.8)), 0:100, ylim = c(100, 0))
}
