% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_minimax}
\alias{sm_shape_minimax}
\title{MiniMax Shape Function}
\usage{
sm_shape_minimax(x, xlim, ascending = TRUE)
}
\arguments{
\item{x}{numeric. Depth vector.}

\item{xlim}{numeric. Parameters: c(depth_min, width_min, depth_max, width_max, skew_min, skew_max).}

\item{ascending}{logical. If TRUE, minimum then maximum. If FALSE, inverted.}
}
\value{
numeric vector in the range 0 to 1 representing the shape.
}
\description{
Models two peaks: minimum then maximum. Useful for complex multi-process soils.
}
\examples{
plot(sm_shape_minimax(0:100, c(30, 10, 70, 15)), 0:100, ylim = c(100, 0))
}
