% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_peak}
\alias{sm_shape_peak}
\title{Peak Shape Function}
\usage{
sm_shape_peak(x, xlim, ascending = TRUE)
}
\arguments{
\item{x}{numeric. Depth vector.}

\item{xlim}{numeric. Parameters: c(depth_of_max, width, skewness). Skewness optional, default 0.}

\item{ascending}{logical. If TRUE, peak. If FALSE, valley.}
}
\value{
numeric vector in the range 0 to 1 representing the shape.
}
\description{
Models a bell-shaped accumulation zone. Useful for Bt horizon, clay peaks.
}
\examples{
plot(sm_shape_peak(0:100, c(50, 20)), 0:100, ylim = c(100, 0))
}
