% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_uniform}
\alias{sm_shape_uniform}
\title{Uniform Shape Function}
\usage{
sm_shape_uniform(x, xlim, ascending = TRUE)
}
\arguments{
\item{x}{numeric. Depth vector.}

\item{xlim}{numeric. Parameters: c(mean_value). The constant value in the range 0 to 1.}

\item{ascending}{logical. If FALSE, inverts the constant value (1 - mean_value).}
}
\value{
numeric vector in the range 0 to 1 representing the shape.
}
\description{
Models a constant value at all depths. Useful for undeveloped or mixed soils.
}
\examples{
plot(sm_shape_uniform(0:100, c(0.5)), 0:100, ylim = c(100, 0))
}
