% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_wetting_front}
\alias{sm_shape_wetting_front}
\title{Wetting Front Shape Function}
\usage{
sm_shape_wetting_front(x, xlim, ascending = TRUE)
}
\arguments{
\item{x}{numeric. Depth vector.}

\item{xlim}{numeric. Parameters: c(depth_inflection_start, depth_inflection_end).}

\item{ascending}{logical. If TRUE, increases with depth. If FALSE, decreases.}
}
\value{
numeric vector in the range 0 to 1 representing the shape.
}
\description{
Models an S-shaped curve; smooth transition between depths. Useful for reaction fronts, leaching.
}
\examples{
plot(sm_shape_wetting_front(0:100, c(20, 50)), 0:100, ylim = c(100, 0))
}
