% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{sm_spline}
\alias{sm_spline}
\title{Mass-preserving Spline Interpolation}
\usage{
sm_spline(x, ...)
}
\arguments{
\item{x}{a data.frame, SoilProfileCollection, or numeric vector}

\item{...}{Additional arguments passed to \code{mpspline2::mpspline()} and/or \code{aqp::spc2mplspline()}}
}
\value{
a data.frame or SoilProfileCollection (when \code{x} is a SoilProfileCollection)
}
\description{
Basic wrapper around \code{mpspline2::mpspline()} and \code{aqp::spc2mpspline()}.
}
\examples{

x <- data.frame(id=1,
                top=c(0,10,20,30,40,50),
                bottom=c(10,20,30,40,50,200),
                prop=c(10,11,22,31,32,31))
plot(sm_spline(x, var_name="prop")[[1]]$est_1cm, 1:200, ylim=c(60,0))

}
