% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sm_shape_sigmoid}
\alias{sm_shape_sigmoid}
\title{Shape functions}
\usage{
sm_shape_sigmoid(x, xlim, ascending = TRUE)
}
\arguments{
\item{x}{numeric. inputs to shape function}

\item{xlim}{numeric. parameter(s) to be applied in shape function with \code{x} as input.}

\item{ascending}{logical. Should the resulting shape be "increasing" (default: \code{TRUE}) or "decreasing"?}
}
\value{
numeric
}
\description{
Shape functions
}
\examples{

plot(sm_shape_sigmoid(0:100, c(20, 50)), 0:100, ylim = c(100, 0))
}
